Source_Code/Back-end_Front-end/Main-Server/requirements.txt
flask
pymysql
flask_restful
flask_jwt_extended
requests
boto3
python-consul
